handle fetching trip data in StatisticPage
