+++ client/src/services/tripService.tsx
@@ -66,7 +66,7 @@ export const addTrip = async (token: string, trip: BaseTrip) => {
   }
 }
 
-export const getAllTrips = async (token: string, year: string, month?: string | null) => {
+export const getAllTrips = async (token: string, year?: string, month?: string | null) => {
   try {
     const authHeader = {"headers": {
       "Authorization": "Bearer " + token
